openvmm_entry: split REPL from VM control into separate modules
Separate the interactive REPL from the VM lifecycle management by
introducing a VmController task and moving the REPL to its own module.

Previously, run_control() interleaved two concerns: owning exclusive VM
resources (worker handles, DiagInspector, vtl2_settings, snapshot config)
and running the interactive command loop. This made it impossible to
reuse either piece independently—e.g., attaching a REPL to an already-
running VM started via ttrpc, or replacing the REPL with a different
control interface.

The VmController task owns all exclusive resources and exposes them
through a mesh RPC channel (VmControllerRpc). The REPL holds cloneable
Sender handles for direct VM interaction (pause/resume/reset, SCSI
hot-plug, etc.) and routes operations needing exclusive access through
the controller. All RPC types derive MeshPayload, so this boundary can
later become cross-process without further changes.

The VmController runs as a spawned task (not a stack future), enabled by
moving VmmMesh ownership into it. This also makes mesh shutdown the
controller's responsibility, which is a better fit since it knows when
all workers have been cleaned up.

This is a stepping stone toward unifying the CLI and ttrpc entry paths
behind a shared VM control interface, and toward supporting dynamic REPL
attach/detach on running VMs.
